+++ spec/support/models/user.rb
@@ -4,6 +4,12 @@ class User
   attr_reader :id
   attr_accessor :name
 
+  class << self
+    def find_by_post_ids(post_ids)
+      post_ids.map { |id| Post.find(id).author }
+    end
+  end
+
   def initialize(id:, name:)
     @id = id
     @name = name